Split views collecting from LaravelDataSource into a new LaravelViewsDataSource.
